requirements.txt
google-generativeai>=0.4.0
python-pptx>=0.6.22
edge-tts>=6.1.9
gTTS>=2.4.0
moviepy>=1.0.3
python-dotenv>=1.0.0
Pillow>=9.0.0
questionary>=2.0.1
requests>=2.31.0
